adding private routes for todos. and testing.

+++ server/server.js
@@ -15,9 +15,10 @@ const app = express();
 
 app.use(bodyParser.json());
 
-app.post("/todos", (req, res) => {
+app.post("/todos", authenticate, (req, res) => {
   var newTodo = new Todo({
-    text: req.body.text
+    text: req.body.text,
+    _creator: req.user._id
   });
   newTodo
     .save()
@@ -25,20 +26,23 @@ app.post("/todos", (req, res) => {
     .catch(err => res.status(400).send(err.message));
 });
 
-app.get("/todos", (req, res) => {
-  Todo.find()
+app.get("/todos", authenticate, (req, res) => {
+  Todo.find({ _creator: req.user._id })
     .then(todos => res.send({ todos }))
     .catch(err => {
       res.status(400).send(err);
     });
 });
 
-app.get("/todos/:todo_id", (req, res) => {
+app.get("/todos/:todo_id", authenticate, (req, res) => {
   let { todo_id } = req.params;
   if (!ObjectID.isValid(todo_id)) {
     return res.status(404).send("Not a valid ID");
   }
-  Todo.findById(todo_id)
+  Todo.findOne({
+    _id: todo_id,
+    _creator: req.user._id
+  })
     .then(todo => {
       if (!todo) {
         return res.status(404).send("Item not found");
@@ -48,12 +52,12 @@ app.get("/todos/:todo_id", (req, res) => {
     .catch(err => res.status(400).send(err.message));
 });
 
-app.delete("/todos/:todo_id", (req, res) => {
+app.delete("/todos/:todo_id", authenticate, (req, res) => {
   let { todo_id } = req.params;
   if (!ObjectID.isValid(todo_id)) {
     return res.status(404).send("Not a valid ID");
   }
-  Todo.findOneAndDelete({ _id: todo_id })
+  Todo.findOneAndDelete({ _id: todo_id, _creator: req.user._id })
     .then(todo => {
       if (!todo) {
         return res.status(404).send("No record with that ID found to delete.");
@@ -64,7 +68,7 @@ app.delete("/todos/:todo_id", (req, res) => {
     .catch(err => res.status(400).send(err.message));
 });
 
-app.patch("/todos/:todo_id", (req, res) => {
+app.patch("/todos/:todo_id", authenticate, (req, res) => {
   let { todo_id } = req.params;
   let body = _.pick(req.body, ["text", "completed"]);
 
@@ -79,7 +83,9 @@ app.patch("/todos/:todo_id", (req, res) => {
     body.completedAt = null;
   }
 
-  Todo.findOneAndUpdate({ _id: todo_id }, body, { new: true })
+  Todo.findOneAndUpdate({ _id: todo_id, _creator: req.user._id }, body, {
+    new: true
+  })
     .then(todo => {
       if (!todo) {
         return res.status(404).send("Todo not found");
